xcm: test, remove mock and bring back testing in xcm_config.rs
