--- runtimes/eden/src/xcm_tests/mod.rs
@@ -1,2 +1,0 @@
-mod conversion;
-mod mock;
